+++ gelidum/collections/__init__.py
@@ -1,7 +1,7 @@
-from gelidum.dependencies import NUMPY_INSTALLED
 from gelidum.collections.frozendict import frozendict  # noqa
 from gelidum.collections.frozenlist import frozenlist  # noqa
 from gelidum.collections.frozenzet import frozenzet  # noqa
+from gelidum.dependencies import NUMPY_INSTALLED
 
 if NUMPY_INSTALLED:
     from gelidum.collections.frozenndarray import frozenndarray  # noqa