+++ gelidum/decorators.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
 import functools
-from typing import Optional, Iterable, Set, Any
-from gelidum import freeze
+from typing import Any, Iterable, Optional, Set
+
+from gelidum.freeze import freeze
 
 
 def freeze_params(params: Optional[Iterable[str]] = None):