+++ gelidum/tests/gelidum_tests/test_freeze_objects.py
@@ -9,168 +9,32 @@ import tempfile
 import threading
 import unittest
 import warnings
-from typing import Dict, List, Union, Any
+from typing import Any, Dict, List, Union
 from unittest.mock import patch
-from gelidum import FrozenException
-from gelidum import freeze
-from gelidum.dependencies import NUMPY_INSTALLED
-from gelidum.collections import frozendict, frozenlist, frozenzet
-from gelidum.frozen import FrozenBase, get_frozen_classes, clear_frozen_classes
+
+from gelidum import FrozenException, freeze
+from gelidum.frozen import FrozenBase, clear_frozen_classes, get_frozen_classes
+from gelidum.tests.gelidum_tests.utils.datetime import utcnow
 
 
-class TestFreeze(unittest.TestCase):
+class TestFreezeObjects(unittest.TestCase):
     def setUp(self) -> None:
         clear_frozen_classes()
 
-    def test_freeze_builtins(self):
-        self.assertEqual(3, freeze(3))
-        self.assertEqual(3.9, freeze(3.9))
-        self.assertEqual(True, freeze(True))
-        self.assertEqual(False, freeze(False))
-        self.assertEqual(None, freeze(None))
-        self.assertEqual(complex(1, 2), freeze(complex(1, 2)))
-        self.assertEqual(b"Bytes", freeze(b"Bytes"))
-        self.assertEqual("String", freeze("String"))
+    def test_freeze_objects_with__call__method(self):
+        class Dummy(object):
+            def __init__(self, value: int):
+                self.attr = value
 
-    def test_freeze_bytearray(self):
-        self.assertEqual(b"Byte array", freeze(bytearray(b"Byte array")))
+            def __call__(self, x: int, y: int):
+                return self.attr + x * y
 
-    def test_freeze_dict(self):
-        frozen_obj: frozendict = freeze({"one": 1, "two": 2})
+        dummy = Dummy(1)
+        frozen_dummy = freeze(dummy, on_freeze="copy", save_original_on_copy=False)
 
-        with self.assertRaises(FrozenException) as context_assignment:
-            frozen_obj["one"] = "another value"  # noqa
+        self.assertEqual(7, frozen_dummy(2, 3))
 
-        with self.assertRaises(FrozenException) as context_clear:
-            frozen_obj.clear()  # noqa
-
-        with self.assertRaises(FrozenException) as context_update:
-            frozen_obj.update({"three": 3})  # noqa
-
-        with self.assertRaises(FrozenException) as context_deletion:
-            del frozen_obj["one"]  # noqa
-
-        self.assertEqual(2, len(frozen_obj))
-        self.assertTrue("one" in frozen_obj)
-        self.assertTrue("two" in frozen_obj)
-        self.assertEqual(1, frozen_obj["one"])
-        self.assertEqual(2, frozen_obj["two"])
-        self.assertEqual("'frozendict' object is immutable", str(context_assignment.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_clear.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_update.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_deletion.exception))
-
-    def test_freeze_list(self):
-        frozen_list = freeze(["one", 2, "three", ["a", "b", "c", 4, 5]])
-        self.assertTrue(isinstance(frozen_list, frozenlist))
-        self.assertEqual(4, len(frozen_list))
-        self.assertEqual("one", frozen_list[0])
-        self.assertEqual(2, frozen_list[1])
-        self.assertEqual("three", frozen_list[2])
-        self.assertTrue(isinstance(frozen_list[3], frozenlist))
-        self.assertEqual(5, len(frozen_list[3]))
-        self.assertEqual("a", frozen_list[3][0])
-        self.assertEqual("b", frozen_list[3][1])
-        self.assertEqual("c", frozen_list[3][2])
-        self.assertEqual(4, frozen_list[3][3])
-        self.assertEqual(5, frozen_list[3][4])
-
-    def test_freeze_list_inplace_true_deprecated_parameter(self):
-        with warnings.catch_warnings(record=True) as caught_warnings:
-            frozen_list = freeze(["one", 2, "three"], inplace=True)
-
-        self.assertTrue(isinstance(frozen_list, frozenlist))
-        self.assertEqual(3, len(frozen_list))
-        self.assertEqual("one", frozen_list[0])
-        self.assertEqual(2, frozen_list[1])
-        self.assertEqual("three", frozen_list[2])
-        self.assertEqual(1, len(caught_warnings))
-        self.assertEqual(
-            "Use of inplace is deprecated and will be removed in next major version (0.6.0)",
-            str(caught_warnings[0].message),
-        )
-
-    def test_freeze_tuple(self):
-        self.assertEqual(("one", 2, "three"), freeze(("one", 2, "three")))
-
-    def test_freeze_set(self):
-        self.assertEqual(frozenzet(["one", 2, "three"]), freeze({"one", 2, "three"}))
-
-    def test_freeze_simple_dataclass(self):
-        from dataclasses import dataclass
-
-        @dataclass
-        class Dummy:
-            attr1: str
-            attr2: str
-            attr3: str = "0"
-
-        dummy = Dummy(attr1="1", attr2="2", attr3="3")
-        frozen_dummy_on_freeze_copy = freeze(dummy, on_freeze="copy")
-        frozen_dummy_inplace = freeze(dummy, on_freeze="inplace")
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy:
-            frozen_dummy_on_freeze_copy.attr1 = "2"
-
-        with self.assertRaises(FrozenException) as context_inplace:
-            frozen_dummy_inplace.attr2 = "2"
-
-        self.assertEqual("Can't assign attribute 'attr1' on immutable instance", str(context_on_freeze_copy.exception))
-        self.assertEqual("Can't assign attribute 'attr2' on immutable instance", str(context_inplace.exception))
-        self.assertEqual(id(dummy), id(frozen_dummy_inplace))
-        self.assertNotEqual(id(dummy), id(frozen_dummy_on_freeze_copy))
-
-    def test_freeze_nested_dataclass(self):
-        from dataclasses import dataclass
-
-        @dataclass
-        class ItemType:
-            id: int
-            name: str
-
-        @dataclass
-        class Item:
-            id: int
-            name: str
-            order: int
-            type: ItemType
-
-        @dataclass
-        class Database:
-            items: List[Item]
-
-        item_type1 = ItemType(1, "item type 1")
-        item_type2 = ItemType(2, "item type 2")
-
-        dummy_database = Database(
-            items=[
-                Item(1, "item 1", 5, item_type1),
-                Item(2, "item 2", 1, item_type1),
-                Item(3, "item 3", 99, item_type2),
-            ]
-        )
-        frozen_dummy_on_freeze_copy = freeze(dummy_database, on_freeze="copy")
-        frozen_dummy_inplace = freeze(dummy_database, on_freeze="inplace")
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy_items_assignment:
-            frozen_dummy_on_freeze_copy.items = []
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy_item_assignment:
-            frozen_dummy_on_freeze_copy.items[0] = Item(9, "item 9", 99, item_type1)
-
-        with self.assertRaises(FrozenException) as context_inplace:
-            frozen_dummy_inplace.attr2 = "2"
-
-        self.assertEqual(
-            "Can't assign attribute 'items' on immutable instance",
-            str(context_on_freeze_copy_items_assignment.exception),
-        )
-        self.assertEqual("'frozenlist' object is immutable", str(context_on_freeze_copy_item_assignment.exception))
-        self.assertEqual("Can't assign attribute 'attr2' on immutable instance", str(context_inplace.exception))
-        self.assertEqual(id(dummy_database), id(frozen_dummy_inplace))
-        self.assertNotEqual(id(dummy_database), id(frozen_dummy_on_freeze_copy))
-
-    def test_freeze_objects_by_copying_and_access_original_object(self):
+    def test_freeze_objects_by_copying_and_accessing_the_original_object(self):
         class DummyChild(object):
             def __init__(self, value: int):
                 self.attr = value
@@ -546,13 +410,13 @@ class TestFreeze(unittest.TestCase):
     @patch("logging.Logger")
     def test_freeze_simple_object_on_update_func_store_update_tries(self, mock_logger):
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self.attr2 = attr2
                 self.attr3 = attr3
 
         class FrozenDummyUpdateTryRecorder(object):
-            def __init__(self, log: logging.Logger):
+            def __init__(self, log: logging.Logger) -> None:
                 self.log = log
                 self.lock = threading.Lock()
                 self.writing_tries: List[Dict] = []
@@ -563,7 +427,7 @@ class TestFreeze(unittest.TestCase):
                 self.original_obj = obj
                 return frozen_object
 
-            def on_update(self, frozen_obj: "FrozenBase", message: str, *args, **kwargs):
+            def on_update(self, frozen_obj: "FrozenBase", message: str, *args, **kwargs) -> None:
                 self.log.warning(message)
                 with self.lock:
                     self.writing_tries.append(
@@ -571,7 +435,7 @@ class TestFreeze(unittest.TestCase):
                             "frozen_obj": frozen_obj,
                             "args": args,
                             "kwargs": kwargs,
-                            "datetime": datetime.datetime.utcnow(),
+                            "datetime": utcnow(),
                         }
                     )
 
@@ -602,9 +466,9 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(id(dummy), id(update_try_recorder.original_obj))
         self.assertEqual(expected_writing_tries, update_try_recorder.writing_tries)
 
-    def test_freeze_simple_object_on_freeze_inplace(self):
+    def test_freeze_simple_object_on_freeze_inplace(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -656,7 +520,7 @@ class TestFreeze(unittest.TestCase):
             "Can't assign attribute '_Dummy__attr3' on immutable instance", str(context_exc_dec_attr3.exception)
         )
 
-    def test_freeze_simple_object_on_freeze_copy(self):
+    def test_freeze_simple_object_on_freeze_copy(self) -> None:
         class Dummy(object):
             def __init__(self, attr1: int, attr2: int, attr3: int):
                 self.attr1 = attr1
@@ -710,15 +574,15 @@ class TestFreeze(unittest.TestCase):
             "Can't assign attribute '_Dummy__attr3' on immutable instance", str(context_exc_dec_attr3.exception)
         )
 
-    def test_freeze_deep_object_on_freeze_inplace(self):
+    def test_freeze_deep_object_on_freeze_inplace(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
 
         class DeepDummy(object):
-            def __init__(self, dummy1: Dummy, dummy2: Dummy, dummy3: Dummy):
+            def __init__(self, dummy1: Dummy, dummy2: Dummy, dummy3: Dummy) -> None:
                 self.dummy1 = dummy1
                 self._dummy2 = dummy2
                 self.__dummy3 = dummy3
@@ -776,15 +640,15 @@ class TestFreeze(unittest.TestCase):
             str(context_exc_setattr_dummy1_attr3.exception),
         )
 
-    def test_freeze_deep_object_on_freeze_copy(self):
+    def test_freeze_deep_object_on_freeze_copy(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
 
         class DeepDummy(object):
-            def __init__(self, dummy1: Dummy, dummy2: Dummy, dummy3: Dummy):
+            def __init__(self, dummy1: Dummy, dummy2: Dummy, dummy3: Dummy) -> None:
                 self.dummy1 = dummy1
                 self._dummy2 = dummy2
                 self.__dummy3 = dummy3
@@ -842,7 +706,7 @@ class TestFreeze(unittest.TestCase):
             str(context_exc_setattr_dummy1_attr3.exception),
         )
 
-    def test_freeze_object_with_file_handler_attribute(self):
+    def test_freeze_object_with_file_handler_attribute(self) -> None:
         with tempfile.TemporaryFile("w") as temp_text_file:
             with tempfile.TemporaryFile("wb") as temp_bin_file:
 
@@ -866,9 +730,9 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual("Text file handlers can't be frozen", str(text_write_context.exception))
         self.assertEqual("Binary file handlers can't be frozen", str(binary_write_context.exception))
 
-    def test_hash_on_freeze_inplace(self):
+    def test_hash_on_freeze_inplace(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -877,9 +741,9 @@ class TestFreeze(unittest.TestCase):
         frozen_dummy = freeze(dummy, on_freeze="inplace")
         self.assertEqual(hash(dummy), hash(frozen_dummy))
 
-    def test_hash_on_freeze_copy(self):
+    def test_hash_on_freeze_copy(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -888,9 +752,9 @@ class TestFreeze(unittest.TestCase):
         frozen_dummy = freeze(dummy, on_freeze="copy")
         self.assertNotEqual(hash(dummy), hash(frozen_dummy))
 
-    def test_hash_frozen_object_as_key_in_dict(self):
+    def test_hash_frozen_object_as_key_in_dict(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -905,7 +769,7 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(hash(dummy), hash(frozen_dummy_inplace))
         self.assertEqual("this is a dummy object", my_dict.get(frozen_dummy_inplace))
 
-    def test_modifying_class_attribute_on_frozen_objects(self):
+    def test_modifying_class_attribute_on_frozen_objects(self) -> None:
         class Dummy(object):
             COUNTER = 1
 
@@ -918,9 +782,9 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(2, frozen_dummy.COUNTER)
         self.assertEqual(2, frozen_dummy_inplace.COUNTER)
 
-    def test_cannot_add_attribute_to_frozen_object(self):
+    def test_cannot_add_attribute_to_frozen_object(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -931,9 +795,9 @@ class TestFreeze(unittest.TestCase):
             frozen_dummy.new_attribute = 99
         self.assertEqual("Can't assign attribute 'new_attribute' on immutable instance", str(context.exception))
 
-    def test_cannot_use_modifying_attribute_method_in_frozen_object(self):
+    def test_cannot_use_modifying_attribute_method_in_frozen_object(self) -> None:
         class Dummy(object):
-            def __init__(self, attr1: int, attr2: int, attr3: int):
+            def __init__(self, attr1: int, attr2: int, attr3: int) -> None:
                 self.attr1 = attr1
                 self._attr2 = attr2
                 self.__attr3 = attr3
@@ -960,14 +824,14 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual("Can't assign attribute '_attr2' on immutable instance", str(context2.exception))
         self.assertEqual("Can't assign attribute '_Dummy__attr3' on immutable instance", str(context3.exception))
 
-    def test_json(self):
+    def test_json(self) -> None:
         data_list = [{"a": 1, "b": 2}, {"a": 11, "b": 22}]
         data_dict = {"first": [{"a": 1}, {"a": 1}]}
 
         self.assertEqual('[{"a": 1, "b": 2}, {"a": 11, "b": 22}]', json.dumps(freeze(data_list)))
         self.assertEqual('{"first": [{"a": 1}, {"a": 1}]}', json.dumps(freeze(data_dict)))
 
-    def test_pickle(self):
+    def test_pickle(self) -> None:
         class DummyForPickle(object):
             def __init__(self, attr: int):
                 self.attr = attr
@@ -985,7 +849,7 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual("Can't assign attribute 'attr' on immutable instance", str(context.exception))
         self.assertEqual(1, frozen_dummy.attr, unpickled_frozen_dummy.attr)
 
-    def test_pickle_object_of_same_name_classes(self):
+    def test_pickle_object_of_same_name_classes(self) -> None:
         import gelidum.tests.gelidum_tests.utils.dummy1
         import gelidum.tests.gelidum_tests.utils.dummy2
 
@@ -1029,7 +893,7 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual("Dummy", frozen_dummy1.get_gelidum_hot_class_name())
         self.assertEqual("Dummy", frozen_dummy2.get_gelidum_hot_class_name())
 
-    def test_hot_class_module_class(self):
+    def test_hot_class_module_class(self) -> None:
         from gelidum.tests.gelidum_tests.utils.dummy1 import Dummy
 
         dummy1 = Dummy(1)
@@ -1046,7 +910,7 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(Dummy.__module__, frozen_dummy2.get_gelidum_hot_class_module())
         self.assertEqual(Dummy.__module__, frozen_dummy3.get_gelidum_hot_class_module())
 
-    def test_hot_class_module_internal_class(self):
+    def test_hot_class_module_internal_class(self) -> None:
         class Dummy(object):
             pass
 
@@ -1064,9 +928,9 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(Dummy.__module__, frozen_dummy2.get_gelidum_hot_class_module())
         self.assertEqual(Dummy.__module__, frozen_dummy3.get_gelidum_hot_class_module())
 
-    def test_count_frozen_classes(self):
+    def test_count_frozen_classes(self) -> None:
         class Dummy(object):
-            def __init__(self, attr: int):
+            def __init__(self, attr: int) -> None:
                 self.attr = attr
 
         dummy1 = Dummy(1)
@@ -1080,7 +944,7 @@ class TestFreeze(unittest.TestCase):
         self.assertSetEqual({frozen_dummy1.__class__}, frozen_classes)
         self.assertEqual(frozen_dummy1.__class__, frozen_dummy2.__class__, frozen_dummy3.__class__)
 
-    def test_invalid_str_for_on_freeze_parameter(self):
+    def test_invalid_str_for_on_freeze_parameter(self) -> None:
         with self.assertRaises(AttributeError) as context:
             freeze(("one", 2, "three"), on_freeze="invalid")
 
@@ -1090,7 +954,7 @@ class TestFreeze(unittest.TestCase):
             str(context.exception),
         )
 
-    def test_invalid_value_for_on_freeze_parameter(self):
+    def test_invalid_value_for_on_freeze_parameter(self) -> None:
         with self.assertRaises(AttributeError) as context:
             freeze(("one", 2, "three"), on_freeze=99)  # noqa
 
@@ -1100,7 +964,7 @@ class TestFreeze(unittest.TestCase):
             str(context.exception),
         )
 
-    def test_invalid_str_for_on_update_parameter(self):
+    def test_invalid_str_for_on_update_parameter(self) -> None:
         with self.assertRaises(AttributeError) as context:
             freeze(("one", 2, "three"), on_update="invalid")
 
@@ -1111,7 +975,7 @@ class TestFreeze(unittest.TestCase):
             str(context.exception),
         )
 
-    def test_invalid_value_for_on_update_parameter(self):
+    def test_invalid_value_for_on_update_parameter(self) -> None:
         with self.assertRaises(AttributeError) as context:
             freeze(("one", 2, "three"), on_update=1)  # noqa
 
@@ -1121,9 +985,9 @@ class TestFreeze(unittest.TestCase):
             str(context.exception),
         )
 
-    def test_structural_sharing_when_freezing_simple_objects(self):
+    def test_structural_sharing_when_freezing_simple_objects(self) -> None:
         class Dummy(object):
-            def __init__(self, value: int):
+            def __init__(self, value: int) -> None:
                 self.value = value
 
         dummy = Dummy(value=1)
@@ -1138,13 +1002,13 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(1, frozen_dummy2.value)
         self.assertEqual(1, frozen_dummy3.value)
 
-    def test_structural_sharing_when_freezing_same_objects_multiple_times(self):
+    def test_structural_sharing_when_freezing_same_objects_multiple_times(self) -> None:
         class DummyAttr(object):
-            def __init__(self, value: str):
+            def __init__(self, value: str) -> None:
                 self.value = value
 
         class Dummy(object):
-            def __init__(self, dummy_attr: FrozenBase):
+            def __init__(self, dummy_attr: FrozenBase) -> None:
                 self.dummy_attr: Union[FrozenBase, DummyAttr] = dummy_attr
 
         frozen_dummy_attr = freeze(DummyAttr("my_value"), on_update="exception", on_freeze="copy")
@@ -1163,13 +1027,13 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(id(frozen_dummy1.dummy_attr), id(frozen_dummy2.dummy_attr))
         self.assertEqual(id(frozen_dummy2.dummy_attr), id(frozen_dummy3.dummy_attr))
 
-    def test_structural_sharing_when_freezing_nested_objects(self):
+    def test_structural_sharing_when_freezing_nested_objects(self) -> None:
         class Id(object):
-            def __init__(self, value: str):
+            def __init__(self, value: str) -> None:
                 self.value = value
 
         class Dummy(object):
-            def __init__(self, value: int, obj_id: Id):
+            def __init__(self, value: int, obj_id: Id) -> None:
                 self.value = value
                 self.obj_id = freeze(obj_id, on_update="exception", on_freeze="copy")
 
@@ -1182,7 +1046,7 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual("unique_id", dummy.obj_id.value)
         self.assertEqual("unique_id", frozen_dummy.obj_id.value)
 
-    def test_structural_sharing_immutable_list(self):
+    def test_structural_sharing_immutable_list(self) -> None:
         class Dummy(object):
             def __init__(self, value: int):
                 self.value = value
@@ -1235,9 +1099,9 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(id(immutable_list_size_3[2]), id(immutable_list_size_6[2]))
         self.assertEqual(id(immutable_list_size_4[3]), id(immutable_list_size_6[3]))
 
-    def test_use_frozen_object_as_dict_key(self):
+    def test_use_frozen_object_as_dict_key(self) -> None:
         class Dummy(object):
-            def __init__(self, value: int):
+            def __init__(self, value: int) -> None:
                 self.attr = value
 
         dummy1 = Dummy(value=1)
@@ -1252,24 +1116,3 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(dummy1, my_dict[frozen_dummy1])
         self.assertTrue(frozen_dummy2 in my_dict)
         self.assertEqual(dummy2, my_dict[frozen_dummy2])
-
-    @unittest.skipUnless(NUMPY_INSTALLED, "numpy is not installed, TestFreeze.test_freeze_ndarray test skipped")
-    def test_freeze_ndarray(self):
-        import numpy as np
-
-        array = np.array([1, 2, 3])
-        frozen_array = freeze(array, on_freeze="copy")
-        frozen_array_copy = frozen_array.copy()
-
-        self.assertEqual(array.shape, frozen_array.shape)
-        self.assertEqual(array.shape, frozen_array_copy.shape)
-        self.assertEqual(array[0], frozen_array[0])
-        self.assertEqual(array[1], frozen_array[1])
-        self.assertEqual(array[2], frozen_array[2])
-        self.assertEqual(array[0], frozen_array_copy[0])
-        self.assertEqual(array[1], frozen_array_copy[1])
-        self.assertEqual(array[2], frozen_array_copy[2])
-        self.assertTrue(isinstance(frozen_array, FrozenBase))
-        self.assertTrue(isinstance(frozen_array, np.ndarray))
-        self.assertTrue(isinstance(frozen_array_copy, FrozenBase))
-        self.assertTrue(isinstance(frozen_array_copy, np.ndarray))