+++ gelidum/tests/gelidum_tests/utils/datetime.py
@@ -1,0 +1,8 @@
+import datetime
+
+
+def utcnow() -> datetime.datetime:
+    try:
+        return datetime.datetime.now(datetime.UTC)
+    except AttributeError:  # pragma: no cover
+        return datetime.datetime.utcnow()  # pragma: no cover