+++ gelidum/tests/gelidum_tests/utils/dummy2.py
@@ -1,3 +1,3 @@
 class Dummy(object):
-    def __init__(self, attr2: int):
+    def __init__(self, attr2: int) -> None:
         self.attr2 = attr2