+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.8.0 (2025-04-05)
+### Features
+- Freezing of functions (i.e. freezing of function attributes).
+### Fixes
+- Split test files.
+- Add isort.
+- datetime.datetime.utcnow is deprecated in Python 3.13.
+- mypy fixes.
+
 ## 0.7.2 (2025-03-30)
 ### Fixes
 - Improve type hinting support.